--- data-access/BASE_URL.ts
@@ -1,1 +1,0 @@
-export const BASE_URL = "https://bootcamp-api.codeit.kr/api/";
