+++ data-access/deleteLink.ts
@@ -1,4 +1,4 @@
-import { axiosInstance } from "./axiosInstance";
+import { axiosInstance } from "./axios/axiosInstance";
 
 export async function deleteLink({ linkId }: { linkId: number }) {
   const response = await axiosInstance.delete(
